+++ test/test-applications/integrations/TestApplication.Wcf.Client.NetFramework/Program.cs
@@ -17,7 +17,6 @@
 using System;
 using System.ServiceModel;
 using System.Threading.Tasks;
-using TestApplication.Wcf.Shared;
 
 namespace TestApplication.Wcf.Client.NetFramework;