+++ test/test-applications/integrations/TestApplication.Wcf.Server.NetFramework/StatusRequest.cs
@@ -1,4 +1,4 @@
-// <copyright file="StatusService.cs" company="OpenTelemetry Authors">
+// <copyright file="StatusRequest.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,23 +14,13 @@
 // limitations under the License.
 // </copyright>
 
-using System;
-using System.ServiceModel;
-using System.Threading.Tasks;
+using System.Runtime.Serialization;
 
 namespace TestApplication.Wcf.Server.NetFramework;
 
-[ServiceBehavior(
-    Namespace = "http://opentelemetry.io/",
-    ConcurrencyMode = ConcurrencyMode.Multiple,
-    InstanceContextMode = InstanceContextMode.Single,
-    UseSynchronizationContext = false,
-    Name = "StatusService")]
-public class StatusService : IStatusServiceContract
+[DataContract]
+public class StatusRequest
 {
-    public Task<StatusResponse> PingAsync(StatusRequest request)
-    {
-        return Task.FromResult(
-            new StatusResponse { ServerTime = DateTimeOffset.UtcNow });
-    }
+    [DataMember]
+    public string Status { get; set; } = string.Empty;
 }